+++ src/domain/challenge/common/tabs/About/AboutSection.tsx
@@ -10,7 +10,7 @@ import {
   EntityDashboardContributors,
   EntityDashboardLeads,
 } from '../../../../community/community/EntityDashboardContributorsSection/Types';
-import { MetricItem } from '../../../../../common/components/composite/common/MetricsPanel/Metrics';
+import { MetricItem } from '../../../../platform/metrics/views/Metrics';
 import PageContentColumn from '../../../../../core/ui/content/PageContentColumn';
 import PageContent from '../../../../../core/ui/content/PageContent';
 import { BlockTitle, Tagline } from '../../../../../core/ui/typography';
