+++ src/domain/challenge/space/DashboardSpaces/SpacesSection.tsx
@@ -7,7 +7,7 @@ import { useSpacesQuery } from '../../../../core/apollo/generated/apollo-hooks';
 import MetricTooltip from '../../../platform/metrics/MetricTooltip';
 import useServerMetadata from '../../../platform/metadata/useServerMetadata';
 import getMetricCount from '../../../platform/metrics/utils/getMetricCount';
-import { MetricItem } from '../../../../common/components/composite/common/MetricsPanel/Metrics';
+import { MetricItem } from '../../../platform/metrics/views/Metrics';
 import { keyBy } from 'lodash';
 import { UserRolesInEntity } from '../../../community/contributor/user/providers/UserProvider/UserRolesInEntity';
 import { Loading } from '../../../../common/components/core';
