--- src/domain/collaboration/callout/creation-dialog/CalloutTemplate/CalloutTemplateProps.ts
@@ -1,5 +1,0 @@
-import { CalloutSumaryProps } from '../../CalloutSummary';
-
-export interface CalloutTemplateProps extends CalloutSumaryProps {
-  onChange?: (templateId: string) => void;
-}
