+++ src/domain/community/contributor/user/constants/translation.constants.ts
@@ -3,6 +3,3 @@ export const ADMIN_TRANSLATION_KEY = 'common.admin';
 export const MEMBER_TRANSLATION_KEY = 'common.member';
 export const LEAD_TRANSLATION_KEY = 'common.lead';
 export const HOST_TRANSLATION_KEY = 'common.host';
-
-export type RoleNameKey = typeof OWNER_TRANSLATION_KEY | typeof ADMIN_TRANSLATION_KEY | typeof MEMBER_TRANSLATION_KEY;
-export type MessageAuthorRoleNameKey = typeof LEAD_TRANSLATION_KEY | typeof HOST_TRANSLATION_KEY;
