--- src/domain/platform/admin/components/index.ts
@@ -1,2 +1,0 @@
-export * from './Group/GroupPage';
-export * from './ListPage';
