--- src/domain/shared/components/LoadingUserCard.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import Grid from '@mui/material/Grid';
-import { UserCard } from '../../../common/components/composite/common/cards';
-
-const LoadingUserCard = () => (
-  <Grid item xs={3}>
-    <UserCard loading />
-  </Grid>
-);
-
-export default LoadingUserCard;
